// src/app/interfaces/anotacao.interface.ts
export interface Anotacao {
  outputData?: any;
  dados?: Anotacao[];
  datFim?: string;
  datIni?: string;
  datNot?: string;
  desNot?: string;
  notFic?: string;
  numCad?: number;
  numEmp?: number;
  staPro?: string;
  tipCol?: number;
  tipNot?: number;
  nomFun?: string;
  motJus?: number;
  tipOpe?: string;
  nomEsc?: string;
  nomLoc?: string;
}